Extract query class for Organization.
