+++ opengever/contact/models/query.py
@@ -1,7 +1,7 @@
 from opengever.contact.models import Contact
 from opengever.contact.models import Organization
 from opengever.contact.models import OrgRole
-from opengever.contact.models.person import Person
+from opengever.contact.models import Person
 from opengever.ogds.models.query import BaseQuery
 from opengever.ogds.models.query import extend_query_with_textfilter
 from sqlalchemy.orm import contains_eager
@@ -30,5 +30,14 @@ class ContactQuery(BaseQuery):
     def get_by_former_contact_id(self, former_contact_id):
         return self.filter_by(former_contact_id=former_contact_id).first()
 
-
 Contact.query_cls = ContactQuery
+
+
+class OrganizationQuery(BaseQuery):
+
+    searchable_fields = ['name']
+
+    def get_by_former_contact_id(self, former_contact_id):
+        return self.filter_by(former_contact_id=former_contact_id).first()
+
+Organization.query_cls = OrganizationQuery
